+++ app/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" name="viewport" content="width=device-width, initial-scale=1">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" type="text/css" href="../dist/main.css">
-    <script async src="../dist/main.js"></script>
+    <script defer src="../dist/main.js"></script>
     <title>Project Mars</title>
   </head>
   <body>
@@ -33,7 +33,12 @@
         <div class="exp__btn"><a class="exp__action">Commencer l'expérience</a></div>
       </div>
       <div class="exp__quizz">
-        <div class="exp__planet planet__three"></div>
+        <div class="exp__planet">
+          <div class="planet__three1"></div>
+          <div class="planet__three2"></div>
+          <div class="planet__three3"></div>
+          <div class="planet__three4"></div>
+        </div>
         <div class="exp__questions">
           <h3 class="exp__questionsTitle">Mars en 2050</h3>
           <div class="exp__questionsText">
