+++ app/styles/main.scss
@@ -79,3 +79,7 @@ body {
 .none {
   display: none !important;
 }
+
+.planet__three1, .planet__three2, .planet__three3, .planet__three4 {
+  height: 100vh;
+}
